### 2019-12-02 #1930 @whedon
[PRE REVIEW]: WEPPCLIFF: A command-line tool to process climate inputs for soil loss models
**Submitting author:** @ryanpmcg (<a href="http://orcid.org/0000-0003-0464-9774">Ryan McGehee</a>)
**Repository:** <a href="https://github.com/ryanpmcg/WEPPCLIFF" target ="_blank">https://github.com/ryanpmcg/WEPPCLIFF</a>
**Version:** 1.4
