### 2019-12-03 9725864
Update AutoTutorial.sh
