+++ TUTORIAL/AutoTutorial.sh
@@ -12,7 +12,7 @@ Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y
 
 Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -fn bpt -la y -verb t
 
-Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y -verb t -o C:/WEPPCLIFF/JUNK -pd t -sdt "2000-01-01 00:00:00" -edt "2004-12-31 24:00:00"
+Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y -verb t -o JUNK -pd t -sdt "2000-01-01 00:00:00" -edt "2004-12-31 24:00:00"
 
 Rscript --vanilla WEPPCLIFF.R --args -f AH537_BPT_CP.csv -cp t -u m PRECIP -sm 2 -la y -ipb t
 
