+++ TUTORIAL/Tutorial.txt
@@ -41,7 +41,7 @@ REQUIREMENTS:
 
 	You must have already installed R (version 3.6.1 or later) and the
 	following directory structure and content must be present for this
-	tutorial to work:
+	tutorial to work (make sure the names are identical):
 
 	C:/ ----+
 		|
@@ -111,7 +111,7 @@ Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv
 Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -fn bpt -la y -verb t
 
 # Change the output location, turn on graphical output, and trim the period analyzed.
-Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y -verb t -o C:/WEPPCLIFF/JUNK -pd t -sdt "2000-01-01 00:00:00" -edt "2004-12-31 24:00:00"
+Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y -verb t -o JUNK -pd t -sdt "2000-01-01 00:00:00" -edt "2004-12-31 24:00:00"
 
 
 ################################################################################
