### 2019-12-05 #1930 @ryanpmcg
Thank you @kbarnhart, I am working on that, but I am new to the community and have not yet built up a network of potential reviewers on GitHub. Am I correct that they should have a GitHub account to be considered as a potential reviewer?
