### 2019-12-05 #1930 @ryanpmcg
There are two post-docs who have now used WEPPCLIFF, and I can ask them to join and review. Otherwise, I saw two persons in the list of people who agreed to be reviewers for the journal (one with climate background and one with soil background).
