### 2019-12-05 #1930 @kbarnhart
@ryanpmcg I think that its most important that the reviewers are well suited to reviewing the software.

If there are people who you think would be good, but who are not yet on GitHub, I can reach out to them via email, see if they are interested in reviewing, and if they are willing, they would make a GitHub account in order to submit the review.
