### 2019-12-06 #1930 @kbarnhart
@ryanpmcg while we are working to find suitable reviewers, I wanted to follow up with an additional comment. Based on looking over the contents of the submission, I would consider adding instructions describing how a user might test that the software has been installed correctly and is working as expected.

Some more information about testing can be found in [this part of the JOSS docs](https://joss.readthedocs.io/en/latest/review_criteria.html#tests).

Given the tutorials you have provided, one possible approach might be a script that when run, executes the tutorial steps, asserts that known outputs are produced, and prints a success message when testing completes successfully.
