+++ AUTOTEST/AutoTest.sh
@@ -1,0 +1,105 @@
+#!/bin/bash
+
+
+################################################################################
+################################  SET DIRECTORY  ###############################
+################################################################################
+
+DIRECTORY=$(cd `dirname $0` && pwd)
+
+cd $DIRECTORY
+
+cd ..
+
+
+################################################################################
+##############################  GLOBAL VARIABLES  ##############################
+################################################################################
+
+TEST=("\n
+	Verifying WEPPCLIFF installation. This test will only take a couple minutes and\n
+	will begin in a few seconds. Please check back momentarily for the results.\n")
+
+CHECKSUM=("\n
+	Running checksum tests. Almost done...\n")
+	 
+SUCCESS=("\n
+	CONGRATULATIONS! Your WEPPCLIFF installation was successfully verified by 8\n
+	automated test runs and 16 highly unique file outputs. Now you are ready to use\n
+	WEPPCLIFF with peace of mind, knowing everything is working as intended.\n")
+
+FAILURE=("\n
+	OOPS! Something is not right. Your WEPPCLIFF installation failed at least one\n
+	of 8 automated test runs. Please remove and reinstall WEPPCLIFF then try this\n
+	test again. If the problem persists and you believe the problem lies within\n
+	WEPPCLIFF, please create an issue at: https://github.com/ryanpmcg/WEPPCLIFF\n
+	\n
+	Thank you! Hopefully the WEPPCLIFF community can help you get up and running\n
+	soon.\n")
+
+CLEARSCREEN=("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n")
+
+
+################################################################################
+###############################  RUN WEPPCLIFF  ################################
+################################################################################
+
+echo -e $CLEARSCREEN
+echo -e $TEST
+sleep 10
+
+Rscript --vanilla WEPPCLIFF.R --args -o $DIRECTORY/OUTPUT -e $DIRECTORY/OUTPUT -fn AutoTest.1 -f ASOS_BPT_KMQE.csv -la y -verb t -ed 1
+Rscript --vanilla WEPPCLIFF.R --args -o $DIRECTORY/OUTPUT -e $DIRECTORY/OUTPUT -fn AutoTest.2 -f AH537_BPT_CP.csv -cp t -u m PRECIP -sm 2 -la y -verb t -ei t -ee ARS -ipb t -pd t -ed 1
+Rscript --vanilla WEPPCLIFF.R --args -o $DIRECTORY/OUTPUT -e $DIRECTORY/OUTPUT -fn AutoTest.3 -f AH537_BPT_IP.csv -cp f -u m PRECIP -sm 2 -la y -verb t -ei t -ee ARS -ipb t -pd t -ed 1
+Rscript --vanilla WEPPCLIFF.R --args -o $DIRECTORY/OUTPUT -e $DIRECTORY/OUTPUT -fn AutoTest.4 -f AH537_1MIN_UP.csv -pi 1 -cp f -u m PRECIP -sm 2 -la y -verb t -ei t -ee ARS -ipb t -pd t -ed 1
+Rscript --vanilla WEPPCLIFF.R --args -o $DIRECTORY/OUTPUT -e $DIRECTORY/OUTPUT -fn AutoTest.5 -f AH537_1MIN.csv -pi 1 -cp f -u m PRECIP -sm 2 -la y -verb t -ei t -ee ARS -ipb t -pd t -ed 1
[…]
