+++ AUTOTEST/README.txt
@@ -1,0 +1,12 @@
+Author: Ryan McGehee
+Created: 9 December 2019
+Purpose: Describe contents of the folder 'TEST'
+
+
+Content:
+
+AutoTest.sh				A shell script to run an automated test on a UNIX system
+
+AutoTest.bat				A batch script to run an automated test on a Windows system
+
+README.txt				This document
