### 2019-12-09 a8a7550
Update Automated Scripts
