### 2019-12-09 c0ed4d1
Update AutoTest.sh
