+++ AUTOTEST/AutoTest.sh
@@ -6,9 +6,7 @@
 ################################################################################
 
 DIRECTORY=$(cd `dirname $0` && pwd)
-
 cd $DIRECTORY
-
 cd ..
 
 
