+++ WEPPCLIFF.R
@@ -472,7 +472,7 @@ create_duration_data = function(data, x) {
     # Convert to datetime (POSIX Count).
     if (length(na.omit(data[[j]][[dt_list[j]]])) == 0) {stop("Datetime data may not be handled correctly. Please check the input file(s) and arguments.")}
     if (length(na.omit(data[[j]][[dt_list[j]]])) > 0) {
-      data[[j]][[dt_list[j]]] = as.POSIXct(as.character(data[[j]][[dt_list[j]]]), format = dtf_list[j])
+      data[[j]][[dt_list[j]]] = as.POSIXct(as.character(data[[j]][[dt_list[j]]]), format = dtf_list[j], tz = "GMT")
       data[[j]] = data[[j]][complete.cases(data[[j]][,1]),]}
     
     # Handle precipitation and non-precipitation breaks.
@@ -505,7 +505,7 @@ create_duration_data = function(data, x) {
   b_rows = intersect(z_rows, s_rows)
   data[[1]] = data[[1]][union(p_rows, b_rows),]
   data[[1]] = data[[1]][order(data[[1]]$DT_1),]
-
+  
   # Modify midnight breakpoint crossings.
   data[[1]] = modify_midnight_breakpoints(df = data[[1]], dtf = dtf1)
   
@@ -514,10 +514,10 @@ create_duration_data = function(data, x) {
 
 # A function to modify breakpoints that cross midnight.
 modify_midnight_breakpoints = function(df, dtf) {
-
+  
   # Calculate time from midnight.
-  df$TFM = sapply(as.POSIXlt(as.character(df$DT_1), format = dtf), function(x) x$hour * 60 + x$min  + x$sec / 60)
-
+  df$TFM = sapply(as.POSIXlt(as.character(df$DT_1), format = dtf, tz = "GMT"), function(x) x$hour * 60 + x$min  + x$sec / 60)
+  
   # Compare time from midnight to break duration.
   dif.vec = df$TFM - df$DUR
   
@@ -666,7 +666,7 @@ run_nonprecip_checks = function(data) {
       
       # Store results.
       data[[i.loc[j]]][[qcnm[j]]] = qcr.vec}}
-
+  
   return(data)}
 
 
@@ -734,7 +734,7 @@ rle_stick_test = function(qcr.vec, stk.th, flag) {
     qcr.starts = qcr.checks - 1
     qcr.ends = qcr.checks + 1
     qcr.identical = which(rle.data$values[qcr.starts] == rle.data$values[qcr.ends])
-
+    
     # Record actual sticks.
     if (length(qcr.identical) > 0) {
       qcr.vec[loc.data[qcr.checks[qcr.identical]]] = flag}}
@@ -774,7 +774,7 @@ remove_qc_fails = function(data) {
       var.vec[qcr.vec != "PASS"] = NA
       data[[i.loc[j]]][[nm[j]]] = var.vec}}
   
-      return(data)}
+  return(data)}
 
 
 # A function to check output daily data quality.
@@ -817,7 +817,7 @@ trim_data = function(data, precip.ts.df, alt.ts.df, d.ts.df, event.list) {
   
   # General Setup
   data = break_storms(data, precip.ts.df, event.list)
-
+  
   # Use all variables for time bounds.
   if (toupper(alt) == "T" & toupper(ipb) == "F"){
     maxs = c(max(data[[precip.ts.df]]$DT_1, na.rm = T), max(data[[alt.ts.df]]$DT_2, na.rm = T), max(data[[d.ts.df]]$DT_3, na.rm = T))
[…]
