+++ WEPPCLIFF.R
@@ -473,6 +473,7 @@ create_duration_data = function(data, x) {
     if (length(na.omit(data[[j]][[dt_list[j]]])) == 0) {stop("Datetime data may not be handled correctly. Please check the input file(s) and arguments.")}
     if (length(na.omit(data[[j]][[dt_list[j]]])) > 0) {
       data[[j]][[dt_list[j]]] = as.POSIXct(as.character(data[[j]][[dt_list[j]]]), format = dtf_list[j])
+      data[[j]][[dt_list[j]]] = as.POSIXct(as.character(data[[j]][[dt_list[j]]]), format = dtf_list[j], tz = "GMT")
       data[[j]] = data[[j]][complete.cases(data[[j]][,1]),]}
     
     # Handle precipitation and non-precipitation breaks.
@@ -517,6 +518,7 @@ modify_midnight_breakpoints = function(df, dtf) {
 
   # Calculate time from midnight.
   df$TFM = sapply(as.POSIXlt(as.character(df$DT_1), format = dtf), function(x) x$hour * 60 + x$min  + x$sec / 60)
+  df$TFM = sapply(as.POSIXlt(as.character(df$DT_1), format = dtf, tz = "GMT"), function(x) x$hour * 60 + x$min  + x$sec / 60)
 
   # Compare time from midnight to break duration.
   dif.vec = df$TFM - df$DUR
@@ -1029,6 +1031,7 @@ start_of_year = function(date) {
   
   # Determine the actual start and setup rounded times.
   real.start = as.POSIXlt(date, format = dtf1)
+  real.start = as.POSIXlt(date, format = dtf1, tz = "GMT")
   good.start = real.start
   month.start = real.start
   day.start = real.start
@@ -1071,6 +1074,7 @@ end_of_year = function(date) {
   
   # Determine the actual end and setup rounded times.
   real.end = as.POSIXlt(date, format = dtf1)
+  real.end = as.POSIXlt(date, format = dtf1, tz = "GMT")
   good.end = real.end
   month.end = real.end
   day.end = real.end
@@ -1194,6 +1198,7 @@ calculate_erosion_indices = function(data, storm.list, storm.df) {
   ke.df = list.cbind(ke.list)
   ei.df = list.cbind(ei.list)
   sp.df = as.data.frame(cbind(format(as.POSIXct(dtimes, origin = "1970-01-01")), depth, dur, pdur, pdur/dur, ip, tp, itw, idw, i30))
+  sp.df = as.data.frame(cbind(format(as.POSIXct(dtimes, origin = "1970-01-01"), tz = "GMT"), depth, dur, pdur, pdur/dur, ip, tp, itw, idw, i30))
   names(sp.df) = c("DT", "DEPTH", "DUR", "PDUR", "PRATIO", "PEAKINT", "PEAKTIME", "TIME_WM_INT", "DEPTH_WM_INT", "I30")
   
   # Calculate storm antecedent times.
@@ -1321,6 +1326,7 @@ process_precip_data = function(data, pcp.in, dly.ts, mly.ts, yly.ts, mly.mn) {
   
   # Remove the first precipitation event if it occurs at midnight.
   first.dt = as.numeric(unlist(as.POSIXlt(data[[pcp.in]]$DT_1[1])))
+  first.dt = as.numeric(unlist(as.POSIXlt(data[[pcp.in]]$DT_1[1], tz = "GMT")))
   first.dur = data[[pcp.in]]$DUR[1]
[…]
