### 2019-12-10 b7eda37
Updated Windows AutoTest
