+++ LIBRARY/mice/NEWS.md
@@ -1,0 +1,481 @@
+---
+title: "News"
+output: github_document
+---
+
+# mice 3.6.0
+
+* Copy `toenail` data from orphaned `DPpackage` package
+* Remove `DPpackage`  from `Suggests` field in `DESCRIPTION` 
+* Adds support for rotated names in `md.pattern()` (#170, #177)
+
+# mice 3.5.0
+
+* This version has some error fixes
+* Fixes a bug in the sampler that ignored imputed values in variables outside the active block (#175, @alexanderrobitzsch)
+* Adds a note to the documenation of `as.mids`() (#173)
+* Removes a superfluous warning from process_mipo() (#92)
+* Fixes an error in the degrees of freedom of the P-value calculation (#171)
+
+# mice 3.4.0 
+
+* Add a hex sticker to the mice package. Designed by Jaden M. Walters.
+* Specify the R3.5.0 random generator in order to pass CRAN tests
+* Remove test-fix.coef.R from tests
+* Adds a rotate.names argument to md.pattern() (#154, #160)
+* Fix to solve the name-matching problem (#156, #149, #147)
+* Fix that removes the pre-check for existence of `mice.impute.xxx()` so that `mice::mice()` works as expected (#55)
+* Solves a bug that crashed `mids2spss()`, thanks Edgar Schoreit (#149)
+* Solves a problem in the routing logic (#149) causing that passive 
+imputation was not done when no predictors were specified. No passive
+imputation correctly will ignore any the specification of 
+`predictorMatrix`.
+* Implements an alternative solution for #93 and #96. Instead of skipping 
+imputation of variables without predictors, `mice 3.3.1` will impute 
+those variables using the intercept only
+* Adds a routine contributed by Simon Grund that checks for deprecated 
+arguments #137
+* Improves the `nelsonaalen()` function for data where variables 
+`time` or `status` have already been defined (#140), thanks matthieu-faron
+
+# mice 3.3.0
+
+* Solves bug in passive imputation (#130). *Warning: This bug may 
+have caused invalid imputations in `mice 3.0.0` - `mice 3.2.0` under 
+passive imputation.*
+* Updates code to `broom 0.5.0` (#128)
+* Solves problem with `mice.impute.2l.norm()` (#129)
+* Use explicit foreign function calls in tests
+
+# mice 3.2.0
+
+* Skip tests for `mice.impute.2l.norm()` (#129)
+* Skip tests for `D1()` (#128)
+* Solve problem with `md.pattern` (#126)
+* Evades warning in `rbind` and `cbind` (#114)
+* Solves `rbind` problem when `method` is a list (#113)
+* More efficient use of `parlmice` (#109)
+* Add `dfcom` argument to `pool()` (#105, #110)
+* Updates to `parlmice` + bugfix (#107)
+
+# mice 3.1.0
+
+* New parallel functionality: `parlmice` (#104)
+* Incorporate suggestion of @JoergMBeyer to `flux` (#102)
+* Replace duplicate code by `estimice` (#101)
+* Better checking for empty methods (#99)
+* Remove problem with `parent.frame` (#98)
+* Set empty method for complete data (#93)
+* Add `NEWS.md`, `index.Rmd` and online package documentation
+* Track `.R` instead of `.r`
+* Patch issue with `updateLog` (#8, @alexanderrobitzsch)
+* Extend README
+* Repair issue `md.pattern` (#90)
+* Repair check on `m` (#89)
+
+# mice 3.0.0 
+         
+Version 3.0 represents a major update that implements the 
+following features: 
+
[…]

[…]
