+++ LIBRARY/EnvStats/scripts/Manual/Chapter09.R
@@ -1,319 +1,0 @@
-#####
-# File:     Chapter09.R
-#
-# Purpose:  Reproduce Examples in Chapter 9 of the book:
-#
-#           Millard, SP. (2013).  EnvStats: An R Package for 
-#             Environmental Statistics.  Springer-Verlag.
-#
-# Author:   Steven P. Millard
-#
-# Last
-# Updated:  2013/03/05
-#####
-
-
-#######################################################################################
-
-library(EnvStats)
-
-############################################################
-
-# 9.3 Monte Carlo Simulation 
-#---------------------------
-
-# 9.3.1 Simulating the Distribution of the Sum of Two Normal Random Variables
-#----------------------------------------------------------------------------
-
-df.100 <- data.frame(simulateMvMatrix(n = 100, seed = 20))
-y.100 <- with(df.100, Var.1 + Var.2)
-
-df.10000 <- data.frame(simulateMvMatrix(n = 10000, seed = 20))
-y.10000 <- with(df.10000, Var.1 + Var.2)
-
-# Table 9.1
-#----------
-summaryStats(y.100)
-
-quantile(y.100, probs = c(0.05, 0.95))
-
-summaryStats(y.10000)
-
-quantile(y.10000, probs = c(0.05, 0.95))
-
-
-# Figure 9.1
-#-----------
-windows()
-hist(y.100, freq = FALSE, col = "cyan", 
-	xlab = expression(paste("Y = ", X[1], " + ", X[2])), 
-	ylab = "Relative Frequency", 
-	main = "Empirical and Theoretical Distributions\nBased on 100 Monte Carlo Trials")
-pdfPlot(param.list = list(mean = 0, sd = sqrt(2)), add = TRUE, pdf.lwd = 3)
-
-
-# Figure 9.2
-#-----------
-windows()
-hist(y.10000, freq = FALSE, breaks = 75, col = "cyan", 
-	xlab = expression(paste("Y = ", X[1], " + ", X[2])), 
-	ylab = "Relative Frequency", 
-	main = "Empirical and Theoretical Distributions\nBased on 10000 Monte Carlo Trials")
-pdfPlot(param.list = list(mean = 0, sd = sqrt(2)), add = TRUE, pdf.lwd = 3)
-
-
-# Clean Up
-#---------
-rm(df.100, y.100, df.10000, y.10000)
-
-#################################################################################
-
-# 9.4.2 Generating Random Numbers from an Arbitrary Distribution
-#---------------------------------------------------------------
-
-# Figure 9.3
-#-----------
-windows()
-cdfPlot(main = paste("Example of Inverse Transformation Method", 
-	"for Standard Normal Distribution", sep = "\n"))
-usr <- par("usr")
-arrows(usr[1], 0.8, x1 = qnorm(0.8), y1 = 0.8, col = "red", lwd = 3)
-arrows(qnorm(0.8), 0.8, x1 = qnorm(0.8), y1 = usr[3], col = "red", lwd = 3)
-
-# Clean Up
-#---------
-rm(usr)
-
-##############################################################################
-
-# 9.4.3 Latin Hypercube Sampling
-#-------------------------------
-
-# Figure 9.4
-#-----------
-windows()
-pdfPlot(curve.fill = FALSE, 
[…]
