+++ LIBRARY/EnvStats/scripts/EPA.2009/Chapter17Examples.R
@@ -1,732 +1,0 @@
-#####
-# File:     Script - EPA09, Chapter 17 Examples.R
-#
-# Purpose:  Reproduce Examples in Chapter 17 of the EPA Guidance Document
-#
-#           USEPA. (2009).  "Statistical Analysis of Ground Water 
-#             Monitoring Data at RCRA Facilities, Unified Guidance."
-#             EPA 530-R-09-007.
-#             Office of Resource Conservation and Recovery, 
-#             Program Information and Implementation Division.
-#             March 2009.
-#
-#           Errata are listed in:
-#           USEPA. (2010).  "Errata Sheet - March 2009 Unified Guidance."
-#             EPA 530/R-09-007a.
-#             Office of Resource Conservation and Recovery, 
-#             Program Information and Implementation Division.
-#             August 9, 2010.
-#
-# Author:   Steven P. Millard
-#
-# Last
-# Updated:  October 03, 2012
-#####
-
-#NOTE:  Unless otherwise noted, differences between what is shown in the
-#       EPA Guidance Document and results shown here are due to the 
-#       EPA Guidance Document rounding intermediate results prior to 
-#       the final result.
-
-library(EnvStats)
-
-#######################################################################################
-
-# Example 17-1, pp. 17-6 to 17-8
-#-------------------------------
-
-# Raw data and summary statistics
-#--------------------------------
-
-head(EPA.09.Ex.17.1.loglead.df)
-#  Month   Well  Well.type LogLead
-#1     1 Well.1 Background    4.06
-#2     2 Well.1 Background    3.99
-#3     3 Well.1 Background    3.40
-#4     4 Well.1 Background    3.83
-#5     1 Well.2 Background    3.83
-#6     2 Well.2 Background    4.34
-
-
-longToWide(EPA.09.Ex.17.1.loglead.df, "LogLead", "Month", "Well", 
-	paste.row.name = TRUE)
-#        Well.1 Well.2 Well.3 Well.4 Well.5 Well.6
-#Month.1   4.06   3.83   4.61   3.53   4.11   4.42
-#Month.2   3.99   4.34   5.14   4.54   4.29   5.21
-#Month.3   3.40   3.47   3.67   4.26   5.50   5.29
-#Month.4   3.83   4.22   3.97   4.42   5.31   5.08
-
-
-summaryStats(LogLead ~ Well, data = EPA.09.Ex.17.1.loglead.df,
-	digits = 2, combine.groups = TRUE, 
-	stats.in.rows = TRUE)
-#       Well.1 Well.2 Well.3 Well.4 Well.5 Well.6 Combined
-#N       4      4      4      4      4      4     24      
-#Mean    3.82   3.96   4.35   4.19   4.8    5      4.35   
-#SD      0.3    0.4    0.66   0.45   0.7    0.4    0.62   
-#Median  3.91   4.03   4.29   4.34   4.8    5.14   4.28   
-#Min     3.4    3.47   3.67   3.53   4.11   4.42   3.4    
-#Max     4.06   4.34   5.14   4.54   5.5    5.29   5.5
-
-summaryStats(LogLead ~ Well.type, data = EPA.09.Ex.17.1.loglead.df,
-	digits = 2, combine.groups = TRUE, 
-	stats.in.rows = TRUE)
-#       Background Compliance Combined
-#N       8         16         24      
-#Mean    3.89       4.58       4.35   
-#SD      0.33       0.61       0.62   
-#Median  3.91       4.48       4.28   
-#Min     3.4        3.53       3.4    
-#Max     4.34       5.5        5.5
-
-
[…]
