+++ LIBRARY/EnvStats/scripts/EPA.2009/Chapter19Examples.R
@@ -1,1113 +1,0 @@
-#####
-# File:     Script - EPA09, Chapter 19 Examples.R
-#
-# Purpose:  Reproduce Examples in Chapter 19 of the EPA Guidance Document
-#
-#           USEPA. (2009).  "Statistical Analysis of Ground Water 
-#             Monitoring Data at RCRA Facilities, Unified Guidance."
-#             EPA 530-R-09-007.
-#             Office of Resource Conservation and Recovery, 
-#             Program Information and Implementation Division.
-#             March 2009.
-#
-#           Errata are listed in:
-#           USEPA. (2010).  "Errata Sheet - March 2009 Unified Guidance."
-#             EPA 530/R-09-007a.
-#             Office of Resource Conservation and Recovery, 
-#             Program Information and Implementation Division.
-#             August 9, 2010.
-#
-# Author:   Steven P. Millard
-#
-# Last
-# Updated:  October 3, 2012
-#####
-
-#NOTE:  Unless otherwise noted, differences between what is shown in the
-#       EPA Guidance Document and results shown here are due to the 
-#       EPA Guidance Document rounding intermediate results prior to 
-#       the final result.
-
-#######################################################################################
-
-library(EnvStats)
-
-# Example 19-1, pp. 19-17 to 19-18
-#---------------------------------
-
-names(EPA.09.Ex.19.1.sulfate.df)
-#[1] "Well"                 "Month"                "Day"                 
-#[4] "Year"                 "Date"                 "Sulfate.mg.per.l"    
-#[7] "log.Sulfate.mg.per.l"
-
-EPA.09.Ex.19.1.sulfate.df[, c("Well", "Date", "Sulfate.mg.per.l", "log.Sulfate.mg.per.l")]
-#    Well       Date Sulfate.mg.per.l log.Sulfate.mg.per.l
-#1  GW-01 1999-07-08             63.0             4.143135
-#2  GW-01 1999-09-12             51.0             3.931826
-#3  GW-01 1999-10-16             60.0             4.094345
-#4  GW-01 1999-11-02             86.0             4.454347
-#5  GW-04 1999-07-09            104.0             4.644391
-#6  GW-04 1999-09-14            102.0             4.624973
-#7  GW-04 1999-10-12             84.0             4.430817
-#8  GW-04 1999-11-15             72.0             4.276666
-#9  GW-08 1997-10-12             31.0             3.433987
-#10 GW-08 1997-11-16             84.0             4.430817
-#11 GW-08 1998-01-28             65.0             4.174387
-#12 GW-08 1999-04-20             41.0             3.713572
-#13 GW-08 2002-06-04             51.8             3.947390
-#14 GW-08 2002-09-16             57.5             4.051785
-#15 GW-08 2002-12-02             66.8             4.201703
-#16 GW-08 2003-03-24             87.1             4.467057
-#17 GW-09 1997-10-16             59.0             4.077537
-#18 GW-09 1998-01-28             85.0             4.442651
-#19 GW-09 1998-04-12             75.0             4.317488
-#20 GW-09 1998-07-12             99.0             4.595120
-#21 GW-09 2000-01-30             75.8             4.328098
-#22 GW-09 2000-04-24             82.5             4.412798
-#23 GW-09 2000-10-24             85.5             4.448516
[…]

[…]
