### 2019-12-10 #1930 @ryanpmcg
@kbarnhart I have uploaded automated testing scripts for OSX and Windows. All changes have been added to an AUTOTEST folder in the repository. I tested both on my two machines, and verified the outputs. I will update the Readme.md with instructions soon; it will be exactly the same as the AutoTutorial shell and batch scripts for consistency.
