### 2019-12-11 #1930 @ryanpmcg
@kbarnhart how should R packages used in WEPPCLIFF be given credit? Right now they are listed in the WEPPCLIFF documentation (acknowledgements). Should they also be listed in the paper somewhere?
