### 2019-12-11 #1930 @kbarnhart
@ryanpmcg

1. I think your approach to adding a testing component to this repository is a good one. Thanks for adding that. Now the repository will be prepared for the reviewers to test the ability to test WEPPCLIFF.

2. If it is possible to cite those packages, I would recommend citing them in the paper. If not, I suspect that your approach thus far is good.

3. I've found one reviewer for this package but want one or two more. If you have recommendations, please let me know.
