### 2019-12-11 #1930 @kbarnhart
@frengers. Thanks for agreeing to review this. The review process is a bit different than other academic journals. The streamlined [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on the software and a [very short paper](https://github.com/openjournals/joss-papers/blob/joss.01887/joss.01887/10.21105.joss.01887.pdf).

This here is a pre-review issue. Once I've attracted sufficient reviewers I'll open a dedicated review issue where check-boxes guide you through the review process.

Once that has occurred, we generally request that reviewers try and complete their review in 2-3 weeks. Obviously many of us are at AGU right now, so I expect that the timeline for getting reviewers and getting reviews may be slightly stretched.

Let me know if you have questions.
