### 2019-12-12 #1930 @ryanpmcg
@kbarnhart in response to your three points:

1. Thank you. The tests are different for each system, and I do not know enough to make them identical for each. The OSX test is the better one (and the system I use most often). The Windows check does not use checksums because there are slight differences in the binary files produced in windows versus OSX from the same R code. So I just check the main output (.cli file) for any text differences with FC. If you or reviewers have suggestions to improve this, I would be glad to hear them.

2. I did not see examples in the 8-10 JOSS papers I read of dependencies being cited in the paper. That is what I based my decision on. I can always add a dependency section to the paper and discuss briefly how each package supports the tool.

3. WEPPCLIFF is the project that linked me to the two potential reviewers, so they may not be impartial in their reviews. Unfortunately, the few people I know that could be suitable reviewers all work with me on various projects. I am sorry that I am not able to help much on this.
