### 2019-12-20 #1930 @kbarnhart
@ryanpmcg I wanted to touch base. I'm having a hard time finding additional reviewers who are able to review this. I suspect that this is in part because AGU was last week, most universities are ending their semester this week or last week, and the end of the year holidays are approaching. I've found one person who is appropriate and willing to review, and at minimum I'll need one more (ideally I try and get 3 reviewers, though sometimes this is not possible).

I want to propose that I take a break from looking for reviewers and restart in early January after people return from vacation/not-working time.

Is this OK with you?

In the meantime, can you look through  [this list of people](https://bit.ly/joss-reviewers) to see if there are any people who look appropriate? Mention them here and put a space between the @ and their GitHub handle (e.g., you'd refer to me as @ kbarnhart).
