### 2019-12-27 2758afc
Update AutoTest.sh
