### 2020-01-02 #1930 @ryanpmcg
@kbarnhart of course that is no problem, I expected things to be put on hold during the holidays. I am sorry for not replying sooner myself.

As for a second reviewer, you may try contacting Andrew.Fullhart@ars.usda.gov (he told me that he would be willing to review the software).
