### 2020-01-15 #1930 @ryanpmcg
@kbarnhart thanks for the update. From the list of people, I have identified a few potentials with justification next to their names below:

@ KristinaRiemer -- lists R as preferred and interest in agriculture
@ dlebauer -- same as above
@ wjakethompson -- R preferred and interest in data management/workflow tools
@ khaors -- R preferred and interest in environmental modeling/hydrogeology
@ ha0ye -- R preferred and interest in time series analysis

WEPPCLIFF is a tool to process general time series climate inputs for agricultural/soil loss models like WEPP. Do you think any of these matches may work?
