### 2020-01-15 #1930 @kbarnhart
:wave: @KristinaRiemer @dlebauer @khaors would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. I'm looking for one more reviewer for this submission.
