### 2020-01-17 #1930 @kbarnhart
:wave: @KristinaRiemer @dlebauer @khaors would you be willing to review this submission for JOSS?
