### 2020-01-17 #1930 @khaors
Dear Katy:

Yes, I am available to review the paper. Please let me know how can I start
the review process.

Thank & Regards,

Oscar

On Fri, Jan 17, 2020, 2:06 PM Katy Barnhart <notifications@github.com>
wrote:

> 👋 @KristinaRiemer <https://github.com/KristinaRiemer> @dlebauer
> <https://github.com/dlebauer> @khaors <https://github.com/khaors> would
> you be willing to review this submission for JOSS?
>
