### 2020-01-17 #2029 @whedon
[REVIEW]: WEPPCLIFF: A command-line tool to process climate inputs for soil loss models
**Submitting author:** @ryanpmcg (<a href="http://orcid.org/0000-0003-0464-9774">Ryan McGehee</a>)
**Repository:** <a href="https://github.com/ryanpmcg/WEPPCLIFF" target ="_blank">https://github.com/ryanpmcg/WEPPCLIFF</a>
**Version:** 1.4

@kbarnhart, @marianadobre, @baharmon please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
