### 2020-01-17 #2029 @kbarnhart
:wave: ryanpmcg, @frengers, @stoor001, @khaors this is the review thread for the paper. All of our communications will happen here from now on.

All reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
