### 2020-02-10 #2029 @kbarnhart
Thanks for the reminder @danielskatz.

@frengers, @stoor001, @khaors could you provide an estimate for when you expect to complete your reviews? We are nearing the end of the 2-4 week window in which we typically try and complete JOSS reviews.

Please let me know if you have any questions or concerns.
