### 2020-02-10 b355550
Corrected date about RIST usage.
