### 2020-02-18 #2029 @frengers
I’m going to try to have this done by the end of the week.

Best,

Francis

> On Feb 17, 2020, at 12:47 PM, Katy Barnhart <notifications@github.com> wrote:
>
> ﻿
> 👋 @frengers, @stoor001, @khaors a friendly reminder to complete your review and/or post and update here. If you have any questions or concerns about the review process, please let me know.
>
> Thanks again for contributing to a JOSS review.
>
