### 2020-03-11 #2029 @kbarnhart
@frengers you mentioned to me earlier today that you were having issues checking the check boxes (this is a problem we are having in general at JOSS right now, sorry).

I believe that you should be able to fix this by clicking the above link. Please let me know if you continue to have checkbox editing issues.
