### 2020-03-23 #2029 @ryanpmcg
@frengers did you run the tutorial successfully first? The autotest.sh script assumes you have done that first. I will revise the language to make this more clear.
