### 2020-03-23 #2029 @ryanpmcg
Adding to my previous comment, I could make installation redundant (i.e. dependencies would be installed in both the autotest scripts and the autotutorial scripts). Right now they are checked in both, but only the tutorial installs them.
