### 2020-03-25 #2029 @kbarnhart
@frengers would you create an issue in the main repository and provide the specific error message you got. I think that a dedicated issue thread would be better to discuss this on rather than this thread. Then @ryanpmcg can respond directly on that thread.

@ryanpmcg I would recommend that installation is separate from running tutorials. I'd also recommend putting installation instructions in the Readme so they are easy to find. If I understand correctly this would just mean putting instructions to navigate to the main directory and execute:
```
Rscript --vanilla WEPPCLIFF.R --args -fr t
```
