### 2020-03-25 #2029 @kbarnhart
@ryanpmcg I also wanted to follow up with an editorial handling update. At this point I'm going to assume that we won't hear from  @stoor001 or @khaors and am going to remove them. I will change from serving as the editor to being the second reviewer (JOSS only requires 2 reviews) and @arfon has agreed to be the handling editor.

Let me know if you have any questions.
