### 2020-03-25 ryanpmcg/WEPPCLIFF#2 @kbarnhart
Installation recommendations
@ryanpmcg this is the first of my review issues for your joss submission.
openjournals/joss-reviews/issues/2029

I was able to install WEPPCLIFF based on the instructions, but have two recommendations:

1. As I mentioned in [this comment](https://github.com/openjournals/joss-reviews/issues/2029#issuecomment-604093133) I would recommend that installation is separate from running tutorials.

I wouldn't recommend installing dependencies as part of the tutorials (though I think that checking for the dependencies) makes sense.

I'd also recommend putting installation instructions in the Readme so they are easy to find. If I understand correctly this would just mean putting instructions to navigate to the main directory and execute:

Rscript --vanilla WEPPCLIFF.R --args -fr t

2. Based on starting through the tutorials and documentation, it seems that it is necessary to run WEPPCLIFF from the directory that WEPPCLIFF.R is located in (please correct me if I'm mistaken). If I were to use WEPPCLIFF on multiple projects this would mean that I'd need to move the file around (which would be non ideal). I'd recommend making an alias for `Rscript --vanilla $WEPP_PATH/WEPPCLIFF.R` so that anywhere on a machine with a successful installation one might execute WEPPCLIFF.
