### 2020-03-25 ryanpmcg/WEPPCLIFF#3 @kbarnhart
Default directories are made when non default directories are specified
Another issue associated with the JOSS review openjournals/joss-reviews/issues/2029

I tried out some of the non default directory names using one of the basic examples.

I renamed INPUTS to SPAM and ran:

```
Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -d SPAM  -o EGGS
```

I was surprised to find that empty directories called INPUT and OUTPUT were still created.
