### 2020-03-25 ryanpmcg/WEPPCLIFF#5 @kbarnhart
Comments on the JOSS paper
A few comments on the JOSS paper as part of the review at  openjournals/joss-reviews/issues/2029.

1. I think that were an acronym occurs right before a citation, the citation format needs to be updated to [ACRONYM, @citation] instead of (ACRONYM) [@citation]. I always get hung up on getting the format right, so know that (a) [this is a page with info on the format](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax), (b) you can recompile the paper with [these Whedon instructions](https://joss.readthedocs.io/en/latest/whedon.html#compiling-papers).

2. I do a reasonable amount of erosion modeling but have never personally used WEPP (yet!?) so I am not particularly familiar with the breakpoint format climate file. I think the first three paragraphs make very clear that there is a real need for software (such as WEPPCLIFF) that support creating climate inputs to WEPP, however, it is only based on guessing after reading the third paragraph that the breakpoint format climate file is the main mechanism of providing climate input to WEPP. I would recommend that you
- [x] State what the main options are for providing climate inputs to WEPP are (perhaps they are just the breakpoint file, perhaps there are others).
- [x] Indicate where a reader can find the specification of this type of input file for WEPP.

3. The figures highlight some of the more advanced qualities of WEPPCLIFF (e.g., quality checking and gap filling) but the text only states that these capabilities exist. I would recommend that you add:
- [x] A small amount of text (1-2 paragraphs max) that highlights these capabilities.
- [x] Expand the figure captions to provide a bit more description.
- [x] Original is not at all visible on Fig 2.
- [x] The difference between original and quality checked is very difficult to see in Figure 3. Perhaps use an unfilled histogram?
