### 2020-03-25 ryanpmcg/WEPPCLIFF#6 @kbarnhart
Improve auto test instructions
https://github.com/openjournals/joss-reviews/issues/2029

My only recommendation would be to be more specific about the path to the test (right now you list path_to_AutoTutorial_here). Instead I would recommend stating something like:

From the top level directory of WEPPCLIFF run bash AUTOTEST/AutoTest.sh (or similar from windows). Otherwise a user would have to guess that the folder is AUTOTEST.
