### 2020-03-26 ryanpmcg/WEPPCLIFF#7 @kbarnhart
All CLI test pass but all RDS tests fail
https://github.com/openjournals/joss-reviews/issues/2029

I ran the autotestst and got the following result:
 Passed CLI Test 1
 Failed RDS Test 1
 Passed CLI Test 2
 Failed RDS Test 2
 Passed CLI Test 3
 Failed RDS Test 3
 Passed CLI Test 4
 Failed RDS Test 4
 Passed CLI Test 5
 Failed RDS Test 5
 Passed CLI Test 6
 Failed RDS Test 6
 Passed CLI Test 7
 Failed RDS Test 7
 Passed CLI Test 8
 Failed RDS Test 8

I'm using R version 3.6.2 (2019-12-12) -- "Dark and Stormy Night" on OS X 10.12.3.

I haven't used R extensively in the last year (and had my computer wiped between now and then), so my R install was new for WEPPCLIFF.

Let me know what I can provide to help diagnose this.
