### 2020-03-26 #2029 @ryanpmcg
Those were all great comments @kbarnhart I will be addressing them all soon. Looking forward to your comments @frengers.
