### 2020-04-03 19e7145
Update AutoTest.sh
