### 2020-04-03 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
I have updated the installation procedure as you recommended. However, I have not implemented aliases before. Do you have a recommendation for how to do this that works for UNIX and Windows systems similarly?
