### 2020-04-03 ryanpmcg/WEPPCLIFF#3 @ryanpmcg
INPUT and OUTPUT are part of a base directory structure created (if they do not exist) in every WEPPCLIFF run. The same is true for the other base directory folders. I can change the behavior to create only the specified folders.
