### 2020-04-03 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
Failing the RDS test is not as important as the CLI test. The CLI test is testing only for human readable differences. The RDS is testing for differences in any internal calculations and those are all stored in binary.

With that said, I still care about the potential differences internally, and to know what is different, I would need the .rds files from your auto test run. They are saved in the auto test folder under OUTPUT.
