### 2020-04-04 49ffc59
permit distinct working and source directories
