+++ WEPPCLIFF.R
@@ -4,7 +4,7 @@
 
 #  Version: 1.4 
 #  Last Updated by: Ryan P. McGehee
-#  Last Updated on: 22 November 2019
+#  Last Updated on: 11 March 2020
 #  Purpose: This program was first designed to create an appropriate input
 #           climate file (.cli) for a WEPP model run. However, the program
 #           has evolved into a much more advanced and capable tool. See the
@@ -36,7 +36,7 @@ ________________________________________________________________________________
   cat(LOGO)}
 
 
-# A function to print a license agreement and accept user response.
+# A function to print a license agreement.
 print_license_agreement = function() {
   LICENSE = {"
   WEPP Climate File Formatter (WEPPCLIFF) Version 1.4
@@ -58,7 +58,6 @@ print_license_agreement = function() {
 
   McGehee, R.P., D.C. Flanagan, and P. Srivastava. 2019. WEPPCLIFF:
       A command-line tool to process climate inputs for soil loss models.
-      Submitted to Journal of Open Source Software on 22 November 2019. 
       Available at: https://github.com/ryanpmcg/WEPPCLIFF
 
   Corresponding Author: R.P. McGehee
@@ -69,7 +68,7 @@ print_license_agreement = function() {
   cat(lr)}
 
 
-# A function to print a license agreement and accept user response.
+# A function to prompt a license agreement and accept user response.
 prompt_license_agreement = function() {
   print_license_agreement()
   cat(lr, lr, "Do you accept these conditions? (Y/N)", lr, lr, collapse = "")
@@ -230,7 +229,7 @@ check_args = function() {
   # Create patterns for grep searching.
   u.patterns = paste("m", "e", collapse = "|")
   tf.patterns = paste("t", "f", collapse = "|")
-  ed.patterns = paste("0", "1", "2", "3", "4", "5", collapse = "|")
+  ed.patterns = paste("0", "1", "2", "3", "4", "5", "6", collapse = "|")
   sm.patterns = paste("1", "2", collapse = "|")
   bf.patterns = paste("1", collapse = "|")
   wi.patterns = paste("0", "1", collapse = "|")
@@ -245,7 +244,7 @@ check_args = function() {
   if (grepl(qc, tf.patterns, ignore.case = T) != T){stop("Invalid Argument: [-qc] must be specified as 't' (true) or 'f' (false).")}
   if (grepl(id, tf.patterns, ignore.case = T) != T){stop("Invalid Argument: [-id] must be specified as 't' (true) or 'f' (false).")}
   if (grepl(pd, tf.patterns, ignore.case = T) != T){stop("Invalid Argument: [-pd] must be specified as 't' (true) or 'f' (false).")}
-  if (grepl(ed, ed.patterns, ignore.case = T) != T){stop("Invalid Argument: [-ed] must be specified as '0' (off), '1' (rds), '2' (json), or '3-5' (precip, daily, or storm csv).")}
+  if (grepl(ed, ed.patterns, ignore.case = T) != T){stop("Invalid Argument: [-ed] must be specified as '0' (off), '1' (rds), '2' (json), '3-5' (precip, daily, or storm csv), or '6' (all).")}
[…]
