### 2020-04-05 ryanpmcg/WEPPCLIFF#3 @kbarnhart
I find the extraneous directories created confusing. For example if a directory called PLOT is created, but no plots are inside (because I didn't specify a run that creates plots, then I'm confused whether I specified the run wrong or whether a plot was supposed to be created but wasn't). A comment that is related to this and #2 is that if you run weppcliff outside of the main directory (I will respond with a recommendation on that on the #2 thread), a LIBRARY directory is created.

I'd recommend only making output directories that are used.

If this is difficult to implement, I'd recommend making very clear in the documentation that (insert list of directory names) is always created.
