### 2020-04-05 ryanpmcg/WEPPCLIFF#2 @kbarnhart
I can't speak directly to working with Windows machines (I presently work on a Mac, a linux, box and a shared HPC platform), but I tried out doing this on a Mac and was successful. I needed to make a few tweaks to the source code (these are on PR #8, you can take or leave them). The key issue that I faced was that I needed to add the main weppcliff directory LIBRARY folder to the R library path (since you've specified the install to put dependencies there and not in the default R location).

I edited my .zshrc file (now that the default OS X shell is zsh instead of bash) to include:

```
WEPPCLIFF_DIR=/pathto/your/directory
alias weppcliff='Rscript --vanilla $WEPPCLIFF_DIR/WEPPCLIFF.R'
```

Once I did that, I was able to copy the input directory into a totally different directory and run tutorial examples. For example:
```
weppcliff --args -f ASOS_BPT_KMQE.csv -la y
```
