### 2020-04-05 ryanpmcg/WEPPCLIFF#7 @kbarnhart
I re-ran the tests today after a fresh install (and without any of the changes I proposed on #8), and now everything fails 😞 . I've attached all the files as a .zip of the OUTPUT folder.

[OUTPUT.zip](https://github.com/ryanpmcg/WEPPCLIFF/files/4435258/OUTPUT.zip)
