+++ WEPPCLIFF.R
@@ -2,9 +2,9 @@
 ############################## PROGRAM METADATA #############################
 #############################################################################
 
-#  Version: 1.4 
+#  Version: 1.4.1
 #  Last Updated by: Ryan P. McGehee
-#  Last Updated on: 11 March 2020
+#  Last Updated on: 5 April 2020
 #  Purpose: This program was first designed to create an appropriate input
 #           climate file (.cli) for a WEPP model run. However, the program
 #           has evolved into a much more advanced and capable tool. See the
@@ -39,7 +39,7 @@ ________________________________________________________________________________
 # A function to print a license agreement.
 print_license_agreement = function() {
   LICENSE = {"
-  WEPP Climate File Formatter (WEPPCLIFF) Version 1.4
+  WEPP Climate File Formatter (WEPPCLIFF) Version 1.4.1
   Copyright (c) 2019 Ryan P. McGehee
 
   This program is free software: you can redistribute it and/or modify
@@ -87,16 +87,15 @@ create_directory = function(var) {
 
 # A function to create the WEPPCLIFF base directory.
 create_base_directory = function() {
-  dirs = c("LIBRARY", "INPUT", "OUTPUT", "EXPORT", "PLOTS", "RUNS", "PROFILES")
+  dirs = c(l, d, o, e, p)
   for (i in dirs){create_directory(i)}}
 
 
 # A function to determine directory location.
 set_wds = function() {
-  home.dir <<- getwd()
-  dir.names = c("lib.dir", "in.dir", "out.dir", "ex.dir", "plot.dir", "run.dir", "prof.dir")
-  dir.locs = c("LIBRARY", "INPUT", "OUTPUT", "EXPORT", "PLOTS", "RUNS", "PROFILES")
-  for (i in 1:length(dir.names)) {assign(dir.names[i], paste(home.dir, "/", dir.locs[i], sep = ""), envir = .GlobalEnv)}}
+  dir.names = c("lib.dir", "in.dir", "out.dir", "ex.dir", "plot.dir")
+  dir.locs = c(l, d, o, e, p)
+  for (i in 1:length(dir.names)) {assign(dir.names[i], dir.locs[i], envir = .GlobalEnv)}}
 
 
 # A function to install R package dependencies on the first run.
@@ -169,10 +168,11 @@ assign_args_to_vars = function() {
 assign_empty_args = function() {
   
   # Execute initial assignment.
-  if (length(d) == 0){assign("d", in.dir, envir = .GlobalEnv)} # Def: current directory
-  if (length(o) == 0){assign("o", out.dir, envir = .GlobalEnv)} # Def: current directory
-  if (length(e) == 0){assign("e", ex.dir, envir = .GlobalEnv)} # Def: current directory
-  if (length(p) == 0){assign("p", plot.dir, envir = .GlobalEnv)} # Def: current directory
+  if (length(d) == 0){assign("d", paste(home.dir, "/INPUT", sep = ""), envir = .GlobalEnv)} # Def: current directory
+  if (length(o) == 0){assign("o", paste(home.dir, "/OUTPUT", sep = ""), envir = .GlobalEnv)} # Def: current directory
+  if (length(e) == 0){assign("e", paste(home.dir, "/EXPORT", sep = ""), envir = .GlobalEnv)} # Def: current directory
[…]
