### 2020-04-06 ryanpmcg/WEPPCLIFF#3 @ryanpmcg
No problem at all. I made the changes, and added the functionality to place the library in a specified location now. The library needed to be specifiable to get past permission issues on some machines. If you find that it works, will you close this issue?
