### 2020-04-06 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
To be clear, I am still not too optimistic about the .rds files. They may never be the same between different operating systems. I never was able to get them identical between OSX and Windows (though I tried to find any differences within).

The .cli test is important, and my plan is to just run the .rds tests quietly while using the .cli tests as the main test.
