### 2020-04-06 ad398c9
Added JOSS Review Suggestions (still working on figures).
