+++ JOSS-PAPER/paper.md
@@ -42,7 +42,7 @@ In short, the erosion modeling community is in need of software that can at leas
 
 After the initial core functions of WEPPCLIFF had been developed, several more advanced functions were added to increase the utility of WEPPCLIFF (and to enhance its ability to deliver high quality climate inputs to WEPP and potentially other models as well). Each of the more impactful features are discussed below with figures when applicable.
 
-Two of these features are more general and are used to refine the time series inputs for all variables supported in WEPPCLIFF (i.e. quality checking and gap filling). WEPPCLIFF quality checking uses a combination of a) hard physical limits, b) adaptive physical rate change limits, and c) adaptive 'spiking', 'streaking', 'sticking', and frequency (return period) analyses to determine potentially faulty or erroneneous data. These analyses are performed for precipitation and non-precipitation variables separately. At the time of publication, a new frequency analysis method based on a generalized extreme value model [climextRemes, @Paciorek:2019] was being incorporated in WEPPCLIFF source code; therefore, that functionality was not active for WEPPCLIFF at publication. Gap filling in WEPPCLIFF is based on a multiple imputation model [MICE, @vanBuuren] which was configured to utilize all available data (within the same time series and from correlated multivariate relationships to other variables) to inform gap-filling of the missing time series. At the time of publication, there was still some need to tweak each of these functions so that they would work seamlessly with the many different data types(temporal resolutions and formats) and diverse climate conditions (severe weather, droughts, nonstationarity, etc.). These functions will be refined as new information becomes available about how to improve them. All other quality checking and gap-filling features were the original work of the WEPPCLIFF lead-developer.
+Two of these features are more general and are used to refine the time series inputs for all variables supported in WEPPCLIFF (i.e. quality checking and gap filling). WEPPCLIFF quality checking uses a combination of a) hard physical limits, b) adaptive physical rate change limits, and c) adaptive 'spiking', 'streaking', 'sticking', and frequency (return period) analyses to determine potentially faulty or erroneneous data. These analyses are performed for precipitation and non-precipitation variables separately. At the time of publication, a new frequency analysis method based on a generalized extreme value model [climextRemes, @Paciorek:2019] was being incorporated in WEPPCLIFF source code; therefore, that functionality was not active for WEPPCLIFF at publication. Gap filling in WEPPCLIFF is based on a multiple imputation model [MICE, @vanBuuren:2011] which was configured to utilize all available data (within the same time series and from correlated multivariate relationships to other variables) to inform gap-filling of the missing time series. At the time of publication, there was still some need to tweak each of these functions so that they would work seamlessly with the many different data types(temporal resolutions and formats) and diverse climate conditions (severe weather, droughts, nonstationarity, etc.). These functions will be refined as new information becomes available about how to improve them. All other quality checking and gap-filling features were the original work of the WEPPCLIFF lead-developer.
 
 ![An example quality checked input timeseries of 15-minute (DSI-3260) precipitation data](KMQE_WIND_VELOCITY_INPUT_TS.png)
 
