### 2020-04-06 ryanpmcg/WEPPCLIFF#9 @kbarnhart
Statement of dependencies
One of the review checklist items in the joss review (openjournals/joss-reviews/issues/2029) states "Is there a clearly-stated list of dependencies?"

The closest thing I found to a list of dependencies is located in the Acknowledgments section of the documentation.

I would recommend listing dependencies somewhere more easily discoverable (perhaps in the readme or in the installation section of the documentation PDF).

A final comment on the issue of installation/dependencies that maybe belongs here but could equally be made on Issues #2 or #3: I would also recommend explicitly stating in the documentation section on installation why the default library location is WEPPCLIFF/LIBRARY instead of into the default R location (you say this on another thread as:

> The library needed to be specifiable to get past permission issues on some machines.

Dependency management and permissions can be a total pain, so I don't think there is any problem with putting the dependencies there, but a comment in the documentation to explain this to an interested/confused individual would be great.
