### 2020-04-06 ryanpmcg/WEPPCLIFF#10 @kbarnhart
Improve community guidelines
Joss review checklist (https://github.com/openjournals/joss-reviews/issues/2029) states:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The Readme in this repo has a contributing section that makes clear that items under (1) and maybe (2) are welcome, but does not indicate how such a request should be made (e.g., an email, and issue, etc).

I'd recommend expanding this section of the Readme by a few sentences to cover all types of contribution listed in the JOSS review checklist, and be specific about mechanisms of communication.
