### 2020-04-06 #2029 @kbarnhart
@ryanpmcg and @arfon I've now opened all the issues I plan to open associated with this review. I expect to be able to complete my review checklist once these issues are resolved. I'll continue to correspond with @ryanpmcg on those issue threads.
