### 2020-04-06 ryanpmcg/WEPPCLIFF#7 @kbarnhart
@ryanpmcg not a problem. I just re-ran them and got the following.
```
 Failed CLI Test 1
 Failed RDS Test 1
 Passed CLI Test 2
 Failed RDS Test 2
 Passed CLI Test 3
 Failed RDS Test 3
 Passed CLI Test 4
 Failed RDS Test 4
 Passed CLI Test 5
 Failed RDS Test 5
 Passed CLI Test 6
 Failed RDS Test 6
 Passed CLI Test 7
 Failed RDS Test 7
 Passed CLI Test 8
 Failed RDS Test 8
```
Mac OS 10.15.2
R version 3.6.2 (2019-12-12) -- "Dark and Stormy Night"
WEPPCLIFF on commit 35de67dd89ca56a66304365375f621563a6199a8

A zip of WEPPCLIFF/AUTOTEST/OUTPUT attached.
[OUTPUT.zip](https://github.com/ryanpmcg/WEPPCLIFF/files/4439662/OUTPUT.zip)

Let me know when/if you want me to try again.
