+++ JOSS-PAPER/paper.md
@@ -44,15 +44,15 @@ After the initial core functions of WEPPCLIFF had been developed, several more a
 
 Two of these features are more general and are used to refine the time series inputs for all variables supported in WEPPCLIFF (i.e. quality checking and gap-filling). WEPPCLIFF quality checking uses a combination of a) hard physical limits, b) adaptive physical rate change limits, and c) adaptive 'spiking', 'streaking', 'sticking', and frequency (return period) analyses to determine potentially faulty or erroneneous data. These analyses are performed for precipitation and non-precipitation variables separately. An example quality checked input using default parameters is provided in Figure 2. At the time of publication, a new frequency analysis method based on a generalized extreme value model [climextRemes, @Paciorek:2019] was being incorporated in WEPPCLIFF source code; therefore, that functionality was not active for WEPPCLIFF at publication.
 
-![An example quality checked input timeseries (daily resolution) of maximum temperature data for a typical midwest climate station using default quality checking parameters.]("11419800 MAXIMUM TEMPERATURE INPUT TS.png")
+![An example quality checked input timeseries (daily resolution) of maximum temperature data for a typical midwest climate station using default quality checking parameters.]("11419800_MAXIMUM_TEMPERATURE_INPUT_TS.png")
 
 Gap-filling in WEPPCLIFF is based on a multiple imputation model [MICE, @vanBuuren:2011] which was configured to utilize all available data (within the same time series and from correlated multivariate relationships to other variables) to inform gap-filling of the missing time series. Figure 3 shows the same quality checked weather station from Figure 2 before and after gap-filling (using default parameters for the 'quick impute' option). At the time of publication, there was still some need to tweak each of these functions so that they would work seamlessly with the many different data types(temporal resolutions and formats) and diverse climate conditions (severe weather, droughts, nonstationarity, etc.). These functions will be refined as new information becomes available about how to improve them. Figures 4 and 5 show examples of how varied data missingness, seasonality, magnitude, variability, etc. can be for any two given stations. All other quality checking and gap-filling features were the original work of the WEPPCLIFF lead-developer.
 
-![An example quality checked output timeseries (daily resolution) of maximum temperature data for the same station using default gap filling parameters for the 'quick impute' option.](11419800 MAXIMUM TEMPERATURE DAILY TS.png)
+![An example quality checked output timeseries (daily resolution) of maximum temperature data for the same station using default gap filling parameters for the 'quick impute' option.](11419800_MAXIMUM_TEMPERATURE_DAILY_TS.png)
 
-![An example gap-filled precipitation time series with heavier, more seasonal rainfall from a longer time period with limited missingness.](KUIL PRECIPITATION DAILY TS.png)
+![An example gap-filled precipitation time series with heavier, more seasonal rainfall from a longer time period with limited missingness.](KUIL_PRECIPITATION_DAILY_TS.png)
 
-![An example gap-filled precipitation time series with lighter, more variable rainfall from a shorter time period with rampant missingness.](KYIP PRECIPITATION DAILY TS.png)
+![An example gap-filled precipitation time series with lighter, more variable rainfall from a shorter time period with rampant missingness.](KYIP_PRECIPITATION_DAILY_TS.png)
 
[…]
