+++ JOSS-PAPER/paper.md
@@ -54,7 +54,7 @@ Gap-filling in WEPPCLIFF is based on a multiple imputation model [MICE, @vanBuur
 
 ![An example gap-filled precipitation time series with lighter, more variable rainfall from a shorter time period with rampant missingness.](KYIP_PRECIPITATION_DAILY_TS.png)
 
-The other two advanced features are specific to soil loss modeling: storm separation and erosion index (EI) calculation. These features provide a critical service from WEPPCLIFF enabling the production of empirical erosivity (more commonly known as R, R-factor, or EI) model inputs from the same time series inputs to physical models. Therefore, the results from WEPPCLIFF can be used to prepare identical climate inputs for empirical and physical models (e.g. WEPP and RUSLE2) so that their various outputs can be compared. The definition of a storm was based on soil loss modeling literature, which utilizes a time and depth threshhold to determine breaks between storms. These are directly specifiable in WEPPCLIFF, and they essentially control how many different storms (with the same I<sub>30\<sub>) should be used to calculate EI. The EI calculation in WEPPCLIFF follows all the rules for various methods in erosion modeling literature, but most importantly, WEPPCLIFF allows the user to specify their own energy equation (a valid R expression based on breakpoint intensity input) for the exploration of new kinetic energy calculation methods and their impact on EI.
+The other two advanced features are specific to soil loss modeling: storm separation and erosion index (EI) calculation. These features provide a critical service from WEPPCLIFF enabling the production of empirical erosivity (more commonly known as R, R-factor, or EI) model inputs from the same time series inputs to physical models. Therefore, the results from WEPPCLIFF can be used to prepare identical climate inputs for empirical and physical models (e.g. WEPP and RUSLE2) so that their various outputs can be compared. The definition of a storm was based on soil loss modeling literature, which utilizes a time and depth threshhold to determine breaks between storms. These are directly specifiable in WEPPCLIFF, and they essentially control how many different storms (with the same I<sub>30</sub>) should be used to calculate EI. The EI calculation in WEPPCLIFF follows all the rules for various methods in erosion modeling literature, but most importantly, WEPPCLIFF allows the user to specify their own energy equation (a valid R expression based on breakpoint intensity input) for the exploration of new kinetic energy calculation methods and their impact on EI.
 
 # Existing Software
 
