### 2020-04-06 ryanpmcg/WEPPCLIFF#9 @ryanpmcg
I added the dependency list to the README. The permission issue is briefly discussed just below the installation command (and informs the user that sometimes permissions prevents the installation if it is not to the C:/ drive directly (or another folder with appropriate permissions).
