### 2020-04-07 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
I found 'an' issue causing differences in the comparisons. I have learned when working with code that an error means exactly the following: there is at least one error.

The files you sent are offset by 2 hours. I am in the Eastern US time zone, and I suppose you are on Mountain time. That would explain why all my testing would not detect it...all the machines were in the same time zone. So all .rds comparisons will fail on account of different dates, and the longer files (the first .rds and .cli test) will result in different calculations from the offset.

I am unsure about the best way to resolve it at the present, and it will take some time for me to learn what to do. I will start searching.
