### 2020-04-08 #2029 @arfon
@baharmon  - many thanks for agreeing to review this submission! Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Also, you'll need to accept this invite to be able to update the checklist as you go: https://github.com/openjournals/joss-reviews/invitations
