### 2020-04-11 ryanpmcg/WEPPCLIFF#2 @baharmon
I have some additional comments regarding installation. I think the installation process and instructions could be better designed as per @kbarnhart suggestions for example. I did a fresh installation of R on Windows and followed your installation instructions, but had some issues with the path and environment variables. I had to install R to the C drive rather than Program Files and use complete file paths to get it to run from C:\WEPPCLIFF.

```
cd C:/WEPPCLIFF
C:/R/R-3.6.3/bin/x64/Rscript --vanilla WEPPCLIFF.R --args -fr t
```

The current installation instructions in the tutorial are only for Windows. Please make a complete UNIX installation guide and tutorial.

I also wanted to note the WEPPCLIFF does not seem to work on Windows Subsystem for Linux.
```
Error in FUN(X[[i]], ...) :
  type 'binary' is not supported on this platform
Calls: main -> install_dependencies -> lapply -> FUN
```
