### 2020-04-11 ryanpmcg/WEPPCLIFF#12 @baharmon
Documentation in Markdown
In addition to the pdf [WEPPCLIFF Documentation](https://github.com/ryanpmcg/WEPPCLIFF/blob/master/WEPPCLIFF%20Documentation%20(v1.4.1).pdf), it would be nice to have more documentation in Markdown linked to the README in or the Wiki. At least a section of the README should point to the documentation pdf. A Manual page in Markdown linked in the README that details the command line tools and arguments (i.e. pages 15-20 of the documentation) would be nice.
