### 2020-04-11 ryanpmcg/WEPPCLIFF#13 @baharmon
Typo in JOSS paper
Intensity is misspelled in:

> The only other related software for providing climate inputs to soil loss models, of which we are aware, is a publicly distributed executable program called Rainfall **Instensity** Summarization Tool [RIST, @USDA:2019].
