### 2020-04-11 #2029 @baharmon
I made several suggestions in the following issues in the WEPPCLIFF repo:
ryanpmcg/WEPPCLIFF#13
ryanpmcg/WEPPCLIFF#12
ryanpmcg/WEPPCLIFF#2
