### 2020-04-19 ryanpmcg/WEPPCLIFF#5 @kbarnhart
@ryanpmcg overall these changes look good. I just have one follow up comment.

Rather than stating "Details on these two input formats can be
found in either CLIGEN, WEPP, or WEPPCLIFF documentation" I would recommend pointing explicitly to where an input specification can be found. I think that this would mean pointing towards Table 1 (Climate input file description (CLIGEN V4.0 format)) from the WEPP user summary (https://www.ars.usda.gov/ARSUserFiles/50201000/WEPP/usersum.pdf).

I think stating this in the caption of Figure 1, and in the section you have the above text (Existing Software) would be sufficient.

My reasoning is that WEPPCLIFF creates a very specific format, and the paper is aimed at a general readership. Thus it is not critical that you explain the format, but it is important that you provide pointers directly to the input specification.
