### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @marianadobre
I was also not able to install WEPPCLIFF on my machine.

![image](https://user-images.githubusercontent.com/37674030/79930686-fd6d8880-83fd-11ea-8365-1ad82fb923ad.png)

![image](https://user-images.githubusercontent.com/37674030/79930701-09594a80-83fe-11ea-8404-c6977e5303ac.png)

I am using version R-3.6.3, which I already had installed on my computer in Program Files and Windows 10.

@baharmon suggested installing R in C and then using complete path

But is an alternative? Can I still run WEPPCLIFF from the Program Files?
