### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
@marianadobre you did not add the correct path for Rscript. You successfully added the path for Rscript in R version 3.6.1. You need to do the same for Rscript in version 3.6.3. Then it should work.
