### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
Can you run Rscript at all (navigate to it’s directory and try to run it; include the full path including the .exe and enclose it all in double quotes because of the space between “Program Files”). If you can’t do that, there is a Windows or R issue that I am not sure I understand.
