### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
@baharmon may be right about that. I might need to change the recommendation to also install R directly on the C drive. Let me know if that works and I will be happy to change it. Also thank you for reviewing the software. I am happy to hear any of your suggestions too.
