### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
@marianadobre I have one last suggestion. Can you try running the AutoTutorial.bat script? It attempts to complete installation automatically for the user. You may have the same issue as before, but maybe not? Will you post the results here if it doesn’t work?
