### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @marianadobre
@ryanpmcg It looks like I can run R.exe but not Rscript.exe. See below:

![image](https://user-images.githubusercontent.com/37674030/79997301-48bf7f80-846e-11ea-84ed-3b011096a070.png)

![image](https://user-images.githubusercontent.com/37674030/79997357-58d75f00-846e-11ea-9cf7-bd790828ebd9.png)

It doesn't work even when I try to lunch the .exe directly (double click on the file) from the /bin directory.  I have three versions of R installed on my computer and this is similar for all three of them. R.exe works but not the Rscript.exe.

Should I next try to install R directly in C?

Also, my pleasure to review the tool. I think it will be quite useful.
