### 2020-04-22 ryanpmcg/WEPPCLIFF#2 @marianadobre
Also, how do I run the AutoTutorial.bat? If I try to save it on my computer it ads .html at the end of file and then takes me to its location in github. And if I delete the .html to have just AutoTutorial.bat, and try to run it, it doesn't execute.
