### 2020-04-22 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
@kbarnhart I haven't forgotten about this. I have had some other pressing business come up that has greater urgency. When I make the changes to the R code, I will ask you to rerun it with those changes. I also plan to incorporate your [pull request](https://github.com/ryanpmcg/WEPPCLIFF/pull/8). Please attach your output when you run it if the .rds comparisons still fail. If I cannot find what is causing the issue after examining the .rds files, I plan to remove the .rds tests because the issue may be OS related.
