### 2020-04-23 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
@marianadobre, actually that message means that Rscript.exe is working. If it did not work it would say command not found or something similar. I do not think you 'need' to install it in the C drive, but that is an option for later.

For now, I would like for you to just navigate to the WEPPCLIFF directory with:
`cd C:/WEPPCLIFF`

then run the AutoTutorial with:
`C:/WEPPCLIFF/TUTORIAL/AutoTutorial.bat`

then just share the results with me
