### 2020-04-23 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
OK, from the same directory try this:
`"C:\Program Files\R\R-3.6.3\bin\Rscript.exe" --vanilla C:\WEPPCLIFF\WEPPCLIFF.R --args -fr t`
