### 2020-04-23 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
Can you close and restart your command prompt then type:
`Rscript`

If that doesn't work try:
`Rscript.exe`
