### 2020-04-23 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
Just curious to see if the PATH variable has the information and just needed to be reset with the new cmd instance.
