### 2020-04-24 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
OK, that means that the path to Rscript is not being saved for some reason. I do not know how to help with that. I just updated the AutoTutorial which now searches for and provides the absolute path for Rscript and the root WEPPCLIFF directory. I do not know if anyone else has suggestions about how to improve installation. I can make an Install.bat script if necessary?
