### 2020-04-24 ryanpmcg/WEPPCLIFF#2 @kbarnhart
This sounds like a windows path issue to me. I don't know whether it is a default part of the R installation on Windows to add the R path to $PATH.

My experience with Windows paths is (very limited as I just started using Windows for work last week) but what I know is that:
- One way to add persistent additions to the $PATH is by modifying the Environment Variables part of System Settings. This typically requires administrator privileges. Lacking administrator settings, believe that setx can be used, though I have not personally used it. (https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/setx).

@ryanpmcg you might consider adding a layer to your install instructions to have a user verify that they can run Rscript before they start trying to use WEPPCLIFF.

Alternatively, (once an appropriate way to set persistent paths in windows that doesn't depend on being an admin and works for both Powershell and command prompt is identified) you might just do something like (for linux/osx)

```
alias weppcliff='$RPATH/Rscript --vanilla $WEPPCLIFF_DIR/WEPPCLIFF.R'
```

I think the windows equivalent is to create a permanent doskey (https://superuser.com/questions/1134368/create-permanent-doskey-in-windows-cmd) I'll try that in the next few days and report back here and/or on PR #8)
