### 2020-04-24 ryanpmcg/WEPPCLIFF#2 @ryanpmcg
If you read back a bit, setx is used in my tutorial.txt. @marianadobre added the path for Rscript under the incorrect directory first, then she added it for the correct directory. I am wondering if the incorrect addition is what caused her first installation attempt not to work.

The AutoTutorial script for Windows provides absolute paths every time for this reason.

@kbarnhart I do want to keep installation as simple as possible. I am convinced that I need to have an automated install procedure that just requires double clicking a .bat script. If you or I figure out the alias for Windows I will have the install procedure set that up automatically, too (for both OS systems).
