### 2020-04-29 10edeb9
Update AutoTest.sh
