### 2020-04-29 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
@kbarnhart I have a new branch with the update. I think I was able to do this on the third attempt. This branch also includes some other updates I plan to roll into the master as well (e.g. alias for UNIX machines, smarter directory path for Windows, should only create directories as needed, etc.)
