### 2020-04-30 55e87a5
Add Automated Installer
