+++ INSTALLER/Install.sh
@@ -1,0 +1,17 @@
+#!/bin/sh
+
+################################################################################
+################################  SET DIRECTORY  ###############################
+################################################################################
+
+DIRECTORY=$(cd `dirname $0` && pwd)
+cd $DIRECTORY
+cd ..
+
+
+################################################################################
+################################  RUN WEPPCLIFF  ###############################
+################################################################################
+
+alias weppcliff="Rscript --vanilla WEPPCLIFF.R --args"
+weppcliff -fr t
