+++ INSTALLER/README.txt
@@ -1,0 +1,12 @@
+Author: Ryan McGehee
+Created: 29 April 2020
+Purpose: Describe contents of the folder 'INSTALLER'
+
+
+Content:
+
+Install.sh				A shell script to install WEPPCLIFF automatically on a UNIX system
+
+Install.bat				A batch script to install WEPPCLIFF automatically on a Windows system
+
+README.txt				This document
