### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
@kbarnhart, I have tested the new script on my Windows machines and now the CLI and RDS files are identical. It was related to the system environment settings and impacting date times, which impacts other things in a time series tool. I merged the branch to the master just to keep it simpler for me. So please test the master branch when available.
