### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @kbarnhart
I did a clean install and got:
```
Passed CLI Test 1
  Failed RDS Test 1

  Passed CLI Test 2
  Failed RDS Test 2

  Passed CLI Test 3
  Failed RDS Test 3

  Passed CLI Test 4
  Failed RDS Test 4

  Passed CLI Test 5
  Failed RDS Test 5

  Passed CLI Test 6
  Failed RDS Test 6

  Passed CLI Test 7
  Failed RDS Test 7

  Passed CLI Test 8
  Failed RDS Test 8
```
This is on a Mac. I haven't gotten to trying things out on my PC...

Contents of WEPPCLIFF/AUTOTEST/OUTPUT attached
[OUTPUT.zip](https://github.com/ryanpmcg/WEPPCLIFF/files/4555892/OUTPUT.zip)
