### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
Well, @kbarnhart, despite the checksum being different, the values stored in at least RDS 2 are exactly the same. I ran my own script in R to check each list and data.frame for differences. Everything checked out. So, I think that this is good enough. I have already update the AutoTest.sh script for UNIX to perform the RDS tests, report results, and still pass the test (if and only if all CLI tests pass). Here is the code I used to check the data frames between three sets (yours, mine, and benchmark):

```
Sys.setenv(TZ = "GMT")

kb.rds = readRDS("/Users/ryanmcgehee/Software/WEPPCLIFF/AUTOTEST/KBOUTPUT/AutoTest.2.rds")
rm.rds = readRDS("/Users/ryanmcgehee/Software/WEPPCLIFF/AUTOTEST/OUTPUT/AutoTest.2.rds")
wc.rds = readRDS("/Users/ryanmcgehee/Software/WEPPCLIFF/AUTOTEST/BENCHMARKS/AutoTest.2.rds")

for (i in c(1,3:5,7,9:14)) {
     kb.df = kb.rds[[i]]
     rm.df = rm.rds[[i]]
     wc.df = wc.rds[[i]]
     print(sum(kb.df == rm.df, na.rm = T))
     print(sum(rm.df == wc.df, na.rm = T))}
```
