### 2020-04-30 6c9ea23
Update AutoTest.sh
