### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
FYI, DFs 2 and 6 are empty for this test, and element 8 is a list, which needs a different test.
