### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
If you run the new AutoTest.sh and think it is good enough, please close this issue.
