+++ TUTORIAL/Tutorial.txt
@@ -5,7 +5,7 @@
 METADATA:
 
 	Author:		Ryan P. McGehee
-	Modified:	26 November 2019
+	Modified:	30 April 2020
 
 PURPOSE:
 
@@ -32,22 +32,26 @@ INSTRUCTIONS:
 
 WARNING:
 
-	You must have all files in the correct directory and you must not alter
+	You must have all files in the correct structure and you must not alter
 	them in any way for the program to execute correctly. For example, Excel
 	changes datetime formats into its default settings if you save a file
 	after opening it. So be careful please!
 
 REQUIREMENTS:
 
-	You must have already installed R (version 3.6.1 or later) and the
-	following directory structure and content must be present for this
-	tutorial to work (make sure the names are identical):
+	You must have already installed R (version 3.6.1 or later) and all of 
+	the WEPPCLIFF dependencies (covered in installation) using default
+	settings. The following directory structure and content must be present
+	for this tutorial to work. If you have opened any of these in another
+	program or have changed the names, it will not work. If you are using
+	Windows, it is recommended for this to be located on the C:/ drive.
 
-	C:/ ----+
-		|
+	
 		+ WEPPCLIFF/ ---+
 				|
 				+ WEPPCLIFF.R
+				|
+				+ LIBRARY/ (all dependencies here)
 				|
 				+ INPUT/ -------+
 						|
@@ -84,18 +88,41 @@ NOTES:
 
 
 ################################################################################
+############################# !!!! IMPORTANT !!!! ##############################
+################################################################################
+
+# All comments are denoted with a '#' at the beginning like this.
+# All Windows-only commands are preceeded with '# WINDOWS:'.
+# All UNIX-only commands are preceeded with '# UNIX:'.
+
+# WINDOWS: start a new command-prompt session.
+# UNIX: start a new terminal session.
+
+
+################################################################################
 ############################# SETUP & INSTALLATION #############################
 ################################################################################
 
-# Adds the path to Rscript for quick access.
+# Installation has already been covered by Install.bat or Install.sh for Windows or UNIX, respectively.
+# This section exists to show you how to perform installation manually if desired.
+
+# Change to the root WEPPCLIFF directory (i.e. cd path\to\WEPPCLIFF).
+# For my Windows machine, this is:
+cd C:\WEPPCLIFF
+
+# WINDOWS: Add the path to Rscript to the system PATH. For my Windows machine running R-3.6.1, this is:
 setx PATH "%PATH%;C:/Program Files/R/R-3.6.1/bin/"
 
-# Changes the directory to the root WEPPCLIFF directory.
-cd C:/WEPPCLIFF
+# WINDOWS: If that worked you need to restart your command prompt, then type: Rscript
+# WINDOWS: If that does not show the screen for Rscript, the PATH variable was not
+# 	   successfully updated, and you will need to use the full path to Rscript
+#	   everytime you see 'Rscript' below.
+
 
 # Installs all WEPPCLIFF dependencies.
 Rscript --vanilla WEPPCLIFF.R --args -fr t
 
+# For slow internet connections, downloads can 'time out' and cause installation to fail.
 # If installation was unsuccessful, you can try the same command again.
 # If installation is still unsuccessful, I do not support non-WEPPCLIFF issues.
 
@@ -155,7 +182,7 @@ Rscript --vanilla WEPPCLIFF.R --args -f AH537_15MIN_UP.csv -pi 15 -cp f -u m PRE
 # Fifteen-minute (0.1 inch precision) data.
 Rscript --vanilla WEPPCLIFF.R --args -f AH537_15MIN_HT.csv -pi 15 -cp f -u m PRECIP -sm 2 -la y -verb t -ei t -ee AH537 -ipb t -pd t -ed 3
 
[…]
