+++ TUTORIAL/Tutorial.md
@@ -1,0 +1,233 @@
+# INTRODUCTION
+
+## METADATA:
+
+	Author:		Ryan P. McGehee
+	Modified:	30 April 2020
+
+## PURPOSE:
+
+	Provide a basic WEPPCLIFF tutorial. This tutorial covers much of but not
+	all WEPPCLIFF capabilities. Read the accompanying documentation for a
+	comprehensive documentation of all functionality, syntax requirements,
+	defaults, and options.
+
+## OVERVIEW:
+
+	This tutorial covers the following topics chronologically:
+
+	+ SETUP & INSTALLATION
+	+ BASIC FUNCTIONALITY
+	+ BREAKPOINT INPUTS
+	+ FIXED INTERVAL INPUTS
+	+ ADVANCED FUNCTIONALITY
+
+## INSTRUCTIONS:
+
+	Read this file for a brief tutorial for using WEPPCLIFF. You may copy
+	the commands directly from this text file into your command prompt or
+	terminal to run the example datasets and learn more about WEPPCLIFF.
+
+## WARNING:
+
+	You must have all files in the correct structure and you must not alter
+	them in any way for the program to execute correctly. For example, Excel
+	changes datetime formats into its default settings if you save a file
+	after opening it. So be careful please!
+
+## REQUIREMENTS:
+
+	You must have already installed R (version 3.6.1 or later) and all of 
+	the WEPPCLIFF dependencies (covered in installation) using default
+	settings. The following directory structure and content must be present
+	for this tutorial to work. If you have opened any of these in another
+	program or have changed the names, it will not work. If you are using
+	Windows, it is recommended for this to be located on the C:/ drive.
+
+	
+		+ WEPPCLIFF/ ---+
+				|
+				+ WEPPCLIFF.R
+				|
+				+ LIBRARY/ (all dependencies here)
+				|
+				+ INPUT/ -------+
+						|
+						+ AH537_1MIN.csv
+						|
+						+ AH537_1MIN_UP.csv
+						|
+						+ AH537_5MIN.csv
+						|
+						+ AH537_10MIN.csv
+						|
+						+ AH537_15MIN.csv
+						|
+						+ AH537_15MIN_HT.csv
+						|
+						+ AH537_15MIN_UP.csv
+						|
+						+ AH537_BPT_CP.csv
+						|
+						+ AH537_BPT_IP.csv
+						|
+						+ ASOS_BPT_KMQE.csv
+
+## NOTES:
+
+	As of WEPPCLIFF Version 1.4 automated gap filling and quality checking
+	routines were determined to be satisfactory for publication, but they
+	are not perfect. In the near future, these routines will be evaluated
+	and published in scientific literature to provide a more complete
+	understanding of their capabilities and limitations.
+
+	KEEP THESE NOTES IN MIND FOR UNDERSTANDING THE CURRENT CAPABILITIES OF
+	WEPPCLIFF AS WELL AS THOSE TO BE INCLUDED IN FUTURE RELEASES.
+
+
+All commands are enclosed in a code box `like this`.
+All Windows-only commands are preceeded with 'WINDOWS:'.
+All UNIX-only commands are preceeded with 'UNIX:'.
+
+WINDOWS: start a new command-prompt session.
+UNIX: start a new terminal session.
+
+
+# SETUP & INSTALLATION
+
+Installation has already been covered by Install.bat or Install.sh for Windows or UNIX, respectively.
+This section exists to show you how to perform installation manually if desired.
+
+Change to the root WEPPCLIFF directory (i.e. cd path\to\WEPPCLIFF).
+For my Windows machine, this is:
+`cd C:\WEPPCLIFF`
+
[…]
