+++ TUTORIAL/Tutorial.md
@@ -93,7 +93,7 @@ WINDOWS: start a new command-prompt session.
 UNIX: start a new terminal session.
 
 
-# SETUP & INSTALLATION
+# SETUP & INSTALLATION:
 
 Installation has already been covered by Install.bat or Install.sh for Windows or UNIX, respectively.
 This section exists to show you how to perform installation manually if desired.
@@ -117,101 +117,93 @@ If installation was unsuccessful, you can try the same command again.
 If installation is still unsuccessful, I do not support non-WEPPCLIFF issues.
 
 
-################################################################################
-############################## BASIC FUNCTIONALITY #############################
-################################################################################
+# BASIC FUNCTIONALITY:
 
-# Run an example input with default settings.
+Run an example input with default settings.
 Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv
 
-# Change the output filename, add early license agreement, and turn on verbosity.
+Change the output filename, add early license agreement, and turn on verbosity.
 Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -fn bpt -la y -verb t
 
-# Change the output location, turn on graphical output, and trim the period analyzed.
+Change the output location, turn on graphical output, and trim the period analyzed.
 Rscript --vanilla WEPPCLIFF.R --args -f ASOS_BPT_KMQE.csv -la y -verb t -o JUNK -pd t -sdt "2000-01-01 00:00:00" -edt "2004-12-31 24:00:00"
 
 
-################################################################################
-############################### BREAKPOINT INPUTS ##############################
-################################################################################
+# BREAKPOINT INPUTS:
 
-# Run an example single-storm, breakpoint input with cumulative precipitation and mixed units (English and metric).
+Run an example single-storm, breakpoint input with cumulative precipitation and mixed units (English and metric).
 Rscript --vanilla WEPPCLIFF.R --args -f AH537_BPT_CP.csv -cp t -u m PRECIP -sm 2 -la y -ipb t
 
-# Run the same input and turn on graphical output, erosion index calculations, and export functionality.
+Run the same input and turn on graphical output, erosion index calculations, and export functionality.
 Rscript --vanilla WEPPCLIFF.R --args -f AH537_BPT_CP.csv -cp t -u m PRECIP -sm 2 -la y -verb t -ei t -ee AH537 -ipb t -pd t -ed 3
 
-# Run the same command using incremental precipitation inputs.
+Run the same command using incremental precipitation inputs.
 Rscript --vanilla WEPPCLIFF.R --args -f AH537_BPT_IP.csv -cp f -u m PRECIP -sm 2 -la y -verb t -ei t -ee AH537 -ipb t -pd t -ed 3
 
 
-################################################################################
-############################# FIXED-INTERVAL INPUTS ############################
-################################################################################
+# FIXED-INTERVAL INPUTS:
 
-# Run several variations of the same command as before using various fixed interval products.
+Run several variations of the same command as before using various fixed interval products.
 
[…]
