### 2020-04-30 3c755f6
Create Manual.md
