+++ Manual.md
@@ -1,0 +1,33 @@
+# WEPPCLIFF MANUAL
+
+## METADATA:
+
+	Author:		Ryan P. McGehee
+	Modified:	30 April 2020
+
+## PURPOSE:
+
+	Provide a manual of WEPPCLIFF syntax, which is easily accessible.
+
+## NOTES:
+
+	WEPPCLIFF is always being improved, which means that its functions and
+	syntax may change over time, but changes are only made when absolutely
+	necessary. Most commands, especially core functionality, will not change
+	much over time.
+	
+	As of WEPPCLIFF Version 1.4 automated gap filling and quality checking
+	routines were determined to be satisfactory for publication, but they
+	are not perfect. In the near future, these routines will be evaluated
+	and published in scientific literature to provide a more complete
+	understanding of their capabilities and limitations.
+
+	KEEP THESE NOTES IN MIND FOR UNDERSTANDING THE CURRENT CAPABILITIES OF
+	WEPPCLIFF AS WELL AS THOSE TO BE INCLUDED IN FUTURE RELEASES.
+
+
+| FLAG | NAME      | DEFAULT | OPTIONS | DESCRIPTION                                                                                                                               |
+|------|-----------|---------|---------|-------------------------------------------------------------------------------------------------------------------------------------------|
+| fr   | first run | F       | T / F   | determines whether or not dependencies should be installed; this should be T for the first run and F for all subsequent runs on a machine |
+|      |           |         |         |                                                                                                                                           |
+|      |           |         |         |                                                                                                                                           |
