### 2020-04-30 b263b50
Update Manual.md
