### 2020-04-30 7e5a205
Update Manual.md
