+++ Manual.md
@@ -25,11 +25,15 @@
 	KEEP THESE NOTES IN MIND FOR UNDERSTANDING THE CURRENT CAPABILITIES OF
 	WEPPCLIFF AS WELL AS THOSE TO BE INCLUDED IN FUTURE RELEASES.
 
-
+## INSTALLATION ARGUMENTS:
 **FLAG**|**NAME**|**DEFAULT**|**OPTIONS**|**DESCRIPTION**
 -----|-----|-----|-----|-----
 fr|first run|F|T / F|determines whether or not dependencies should be installed; this should be T for the first run and F for all subsequent runs on a machine
 la|license agreement| |Y / N|the default is to not provide a response to the license agreement
+
+## INPUT/OUTPUT ARGUMENTS:
+**FLAG**|**NAME**|**DEFAULT**|**OPTIONS**|**DESCRIPTION**
+-----|-----|-----|-----|-----
 d|input directory|./INPUT|a directory|the parent directory of the single input file or the directory of files to process; the WEPPCLIFF home directory is the default local root
 o|output directory|./OUTPUT|a directory|the location where all .cli files will be written; the WEPPCLIFF home directory is the default local root
 e|export directory|./EXPORT|a directory|the location where all binary .rds (binary) or .json (text) files will be written; the WEPPCLIFF home directory is the default local root
@@ -38,6 +42,10 @@ l|library directory|./LIBRARY|a directory|the location where all WEPPCLIFF depen
 f|filename|NONE|.txt, .csv, or .tsv file|the input file if there is only one; leaving this blank will cause the program to try and process all files in the input directory; accepts tab or comma separated entries (must be .tsv or .csv, respectively) or you can specify a delimiter for .txt files with -delim
 fn|output filename|station|a string|the name which will be used as default for all files written during WEPPCLIFF processing; specifying 'station' will search the input data for a station name to be assigned to the outputs (especially useful for parallel processing)
 delim|file delimiter|a single space|a string|specifies a delimiter to use when the input file is of type .txt
+
+## FUNCTIONALITY ARGUMENTS:
+**FLAG**|**NAME**|**DEFAULT**|**OPTIONS**|**DESCRIPTION**
+-----|-----|-----|-----|-----
 u|units|M|M <vars> / E <vars>|the units of the input file (Metric or English) followed by variables which are exceptions must be specified in order for the correct conversions to be applied
 qc|quality control|F|T / F|determines whether or not to perform quality control routines for the input data; the flags -qcop, -qcth, and -qcdf control the methodology
 id|impute missing data|F|T / F|determines whether or not to engage a multiple imputation model to fill missing data; the flags -im, -io, -qi, and -iv control the methodology
@@ -45,24 +53,44 @@ pd|plot data|F|T / F|determines whether or not to create graphical output for bo
 ed|export data|0|0 / 1 / 2 / 3 / 4 / 5|determines whether or not to generate binary or human readible, output of internal calculations; (0) no files written; (1) .rds binary files written; (2) .json text files written; (3-5) .csv text files written with precipitation, daily, or storm timeseries dataframes, respectively
 alt|alternative data|F|T / F|specifies whether or not to use alternative data to calculate minimum, maximum, and dew point temperatures; this requires the variables AIR\_TEMP and REL\_HUM
 pmd|preserve missing data|F|T / F|specifies whether or not to maintain NA values throughout execution; when true, this converts all non-numeric inputs and calculations and the WEPPCLIFF missing data value (-99999) values to NA; internal calculations and plots are affected according to various conventions throughout execution
+
+## PRECIPITATION ARGUMENTS:
+**FLAG**|**NAME**|**DEFAULT**|**OPTIONS**|**DESCRIPTION**
+-----|-----|-----|-----|-----
[…]
