### 2020-04-30 ryanpmcg/WEPPCLIFF#12 @ryanpmcg
@baharmon please see the updated installation, testing, tutorial, and manual content in the README. Here are some links to new markdown content:
[Manual](https://github.com/ryanpmcg/WEPPCLIFF/blob/master/Manual.md)
[Tutorial](https://github.com/ryanpmcg/WEPPCLIFF/blob/master/TUTORIAL/Tutorial.md)
