### 2020-04-30 ryanpmcg/WEPPCLIFF#7 @kbarnhart
@ryanpmcg my thought is that I would be confused by an unexplained exit code. If the RDS tests aren't going to influence whether the tests pass or fail, I'd comment them out. Or at minimum state that an exit code of 0 on RDS means that the binary files are not the same but that is know to be based on system differences.

You could also run the above script as the RDS test since it ultimately demonstrates what you care about (that the contents of the RDS are the same).

Whatever the case, I think it is fair to say that the CLI tests are sufficient as tests.

@thatmushroom (my local R expert) thought that there might be a difference between the system md5 and the internal R md5 and that running md5 within R might solve the issue... In fact there was no difference.
```
(base) barnhark@Katherines-MacBook-Pro WEPPCLIFF % md5 AUTOTEST/OUTPUT/AutoTest.1.rds
MD5 (AUTOTEST/OUTPUT/AutoTest.1.rds) = 367cdb2d4f48f81ce3a3f2cd326dc9ec
(base) barnhark@Katherines-MacBook-Pro WEPPCLIFF % md5 AUTOTEST/BENCHMARKS/AutoTest.1.rds
MD5 (AUTOTEST/BENCHMARKS/AutoTest.1.rds) = 3ccedaa3a5345579b959d7ab50d42f04
(base) barnhark@Katherines-MacBook-Pro WEPPCLIFF % R -q -e "library('tools'); md5sum('AUTOTEST/OUTPUT/AutoTest.1.rds')"
> library('tools'); md5sum('AUTOTEST/OUTPUT/AutoTest.1.rds')
    AUTOTEST/OUTPUT/AutoTest.1.rds
"367cdb2d4f48f81ce3a3f2cd326dc9ec"
>
>
(base) barnhark@Katherines-MacBook-Pro WEPPCLIFF % R -q -e "library('tools'); md5sum('AUTOTEST/BENCHMARKS/AutoTest.1.rds')"
> library('tools'); md5sum('AUTOTEST/BENCHMARKS/AutoTest.1.rds')
AUTOTEST/BENCHMARKS/AutoTest.1.rds
"3ccedaa3a5345579b959d7ab50d42f04"
>
>
(base) barnhark@Katherines-MacBook-Pro WEPPCLIFF %
```
