### 2020-05-01 ryanpmcg/WEPPCLIFF#7 @ryanpmcg
@kbarnhart I will remove the exit codes and update the script and close this issue.
