### 2020-05-02 187b55f
Update AutoTest.sh
