### 2020-05-11 #2029 @arfon
@ryanpmcg - could you take a look at the list of suggested DOIs for your references and add them to your BibTeX as entries like this `doi = {10.18637/jss.v045.i03}` if they are correct?
