### 2020-05-11 #2029 @arfon
@ryanpmcg - Once you've addressed the DOI question above, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
