### 2020-05-11 7b63695
Added missing doi's
