+++ JOSS-PAPER/paper.bib
@@ -24,6 +24,7 @@
 	Volume = {45},
 	Number = {3},
 	Pages = {1-67},
+	doi = {10.18637/jss.v045.i03},
 	Url = {https://www.jstatsoft.org/v45/i03/}}
   
   @webpage{USDA:2019,
@@ -66,6 +67,7 @@
 	Pages = {334-349},
 	Title = {Updated climate database and impacts on WEPP model predictions},
 	Volume = {74},
+	doi = {10.2489/jswc.74.4.334},
 	Year = {2019}}
 
 @book{USDA:1995,
@@ -94,6 +96,7 @@
 	Title = {Erosivity index values for northern Mississippi},
 	Volume = {38},
 	Year = {1995},
+	doi = {10.13031/2013.27921},
 	Bdsk-File-1 = {YnBsaXN0MDDSAQIDBFxyZWxhdGl2ZVBhdGhZYWxpYXNEYXRhXxBtLi4vRG9jdW1lbnRzL1Jlc2VhcmNoL1Byb2plY3RzL0FVIFRoZXNpcy9MaXRlcmF0dXJlIFJldmlldy9Fcm9zaXZpdHkvRUkvRUkgVmFsdWVzIGluIE5vcnRoZXJuIE1pc3Npc3NpcHBpLnBkZk8RAhwAAAAAAhwAAgAADE1hY2ludG9zaCBIRAAAAAAAAAAAAAAAAAAAAAAAAABCRAAB/////x9FSSBWYWx1ZXMgaW4gTm9ydGgjRkZGRkZGRkYucGRmAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AAAAAAAAAAAAAAAAAAEACAAACiBjdQAAAAAAAAAAAAAAAAACRUkAAgB+LzpVc2VyczpyeWFubWNnZWhlZTpEb2N1bWVudHM6UmVzZWFyY2g6UHJvamVjdHM6QVUgVGhlc2lzOkxpdGVyYXR1cmUgUmV2aWV3OkVyb3Npdml0eTpFSTpFSSBWYWx1ZXMgaW4gTm9ydGhlcm4gTWlzc2lzc2lwcGkucGRmAA4ATAAlAEUASQAgAFYAYQBsAHUAZQBzACAAaQBuACAATgBvAHIAdABoAGUAcgBuACAATQBpAHMAcwBpAHMAcwBpAHAAcABpAC4AcABkAGYADwAaAAwATQBhAGMAaQBuAHQAbwBzAGgAIABIAEQAEgB8VXNlcnMvcnlhbm1jZ2VoZWUvRG9jdW1lbnRzL1Jlc2VhcmNoL1Byb2plY3RzL0FVIFRoZXNpcy9MaXRlcmF0dXJlIFJldmlldy9Fcm9zaXZpdHkvRUkvRUkgVmFsdWVzIGluIE5vcnRoZXJuIE1pc3Npc3NpcHBpLnBkZgATAAEvAAAVAAIAEv//AAAACAANABoAJACUAAAAAAAAAgEAAAAAAAAABQAAAAAAAAAAAAAAAAAAArQ=}}
 
 @article{McGehee:2018,
@@ -108,6 +111,7 @@
 	Title = {Benchmarking reliable erosion indices from quarter-hour station data for climate studies in the Southeastern United States},
 	Volume = {73},
 	Year = {2018},
[…]
